revert back to non-pb sort using shorts

+++ core/src/main/scala/org/apache/spark/network/netty/protocol.scala
@@ -270,8 +270,8 @@ private[netty] object ProtocolUtils {
 
     if (b == 1) {
       val shuffleId = in.readByte()
-      val mapId = in.readInt()
-      val reduceId = in.readInt()
+      val mapId = in.readShort()
+      val reduceId = in.readShort()
       new ShuffleBlockId(shuffleId, mapId, reduceId)
     } else if (b == 2) {
       val broadcastId = in.readLong()
